♻️ Refactor: 공통 로직 useFetchProduct 훅으로 분리
